chore(cli): Phase 4 — decompose plugin-runtime.ts into a folder
Split the ~770-line runtime/plugin-runtime.ts into focused modules
under runtime/plugin-runtime/:

- errors.ts: formatPluginError + guardHook helper that wraps the
  per-hook try/catch boilerplate previously inlined 11 times
- loader.ts: normalizePluginRegistration, validatePlugin,
  importPluginModule, parseCliMajor
- executor-debug.ts: wrapExecutorWithDebug. A small trace() helper
  collapses the repeated start/OK/FAILED debug pattern that used
  to be copy-pasted across 9 executor methods
- null-executor.ts: the NULL_EXECUTOR singleton
- hooks.ts: each runOnXxx as a standalone exported function taking
  loaded plugins explicitly (UNFILTERED_TABLE_SCOPE lives here)
- index.ts: loadPluginRuntime assembles the PluginRuntime by
  delegating to the modules above

No behaviour change. Importers updated to use plugin-runtime/index.js.
