+++ packages/cli/src/bin/chkit.ts
@@ -15,7 +15,7 @@ import { cmdInit } from '../commands/init.js'
 import { loadConfig } from '../runtime/config.js'
 import { GLOBAL_FLAGS } from '../runtime/global-flags.js'
 import { formatGlobalHelp, formatCommandHelp } from '../runtime/help.js'
-import { loadPluginRuntime } from '../runtime/plugin-runtime.js'
+import { loadPluginRuntime } from '../runtime/plugin-runtime/index.js'
 import { getInternalPlugins } from '../internal-plugins/index.js'
 import { CLI_VERSION } from '../runtime/version.js'
 import { debug } from '../runtime/debug.js'
